+++ ConsomiTounsi/src/main/java/tn/esprit/spring/controller/Command_lineController.java
@@ -23,7 +23,7 @@ public class Command_lineController{
 	tn.esprit.spring.Repository.Command_lineRepository Command_lineRepository;
 	//http://localhost:8081/ConsomiTounsi/servlet/
 		@PostMapping("/addCommand_line")
-		public int addCommand_line(@RequestBody Command_line command) {
+		public long addCommand_line(@RequestBody Command_line command) {
 			command_lineService.addCommandline(command);
 			//Command_lineRepository.getQuntityIngredient(0);
 			return command.getId();
@@ -48,13 +48,13 @@ public class Command_lineController{
 		
 
 	    @PutMapping(value = "/affecterProductACommand_line/{id}/{id}") 
-		public void affecterProductACommand_line(@PathVariable("id")int id, @PathVariable("id")int idc) {
+		public void affecterProductACommand_line(@PathVariable("id")long id, @PathVariable("id")long idc) {
 			command_lineService.affecterPoductACommand_line(id, idc);
 			command_lineService.update_prix();
 		}
 	    
 	    @PutMapping(value = "/affecterCommandACommand_line/{reference}/{id}") 
-			public void affecterCommandACommand_line(@PathVariable("reference")int reference, @PathVariable("id")int id) {
+			public void affecterCommandACommand_line(@PathVariable("reference")long reference, @PathVariable("id")long id) {
 				command_lineService.affecterCommandACommand_line(reference, id);
 				
 			}
@@ -71,7 +71,7 @@ public class Command_lineController{
 		}*/
 	    
 		@DeleteMapping(value = "/deleteByidc/{id}")
-		public void deleteById(@PathVariable("id")int id) {
+		public void deleteById(@PathVariable("id")long id) {
 			command_lineService.deleteById(id);
 		}
 			  @PutMapping(value = "/updatecommande_line")
